backend/list_gemini_models.py
import os

from dotenv import load_dotenv
import google.generativeai as genai

load_dotenv()

api_key = os.getenv("GEMINI_API_KEY")
if not api_key:
    print("No API key found")
else:
    genai.configure(api_key=api_key)
    print("Listing available Gemini models...")
    try:
        for model in genai.list_models():
            if "generateContent" in model.supported_generation_methods:
                print(f"Name: {model.name}")
    except Exception as e:
        print(f"Error listing models: {e}")
